+++ CHANGELOG.md
@@ -9,6 +9,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Blocks(`@appsemble/data-loader`): Add `skipInitialLoad` parameter.
 - Blocks(`@appsemble/map`): Make `move` emit event optional. When omitted, the map block won’t emit
   refresh events.