+++ CHANGELOG.md
@@ -9,10 +9,13 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Blocks(`@appsemble/data-loader`): Add `skipInitialLoad` parameter.
 - Blocks(`@appsemble/map`): Add marker clustering.
 - Blocks(`@appsemble/map`): Add `disableClustering` parameter.
 - Blocks(`@appsemble/map`): Add `maxClusterRadius` parameter. This can be used to determine when the
   map should cluster markers.
+- Blocks(`@appsemble/map`): Make `move` emit event optional. When omitted, the map block won’t emit
+  refresh events.
 
 ### Fixed